chore: refactor useNonEvmTokens into a hook
