+++ app/components/UI/Bridge/hooks/useNonEvmTokens/index.ts
@@ -1,0 +1,1 @@
+export { useNonEvmTokens } from './useNonEvmTokens';
